REPO-3494 : retain the history on the cancelled workflows adding property system.workflow.engine.activiti.retentionHistoricProcessInstance=true

+++ src/main/java/org/alfresco/repo/workflow/activiti/ActivitiWorkflowEngine.java
@@ -264,7 +264,10 @@ public class ActivitiWorkflowEngine extends BPMEngine implements WorkflowEngine
             WorkflowInstance result =  typeConverter.convert(deletedInstance);
             
             // Delete the historic process instance
-            historyService.deleteHistoricProcessInstance(deletedInstance.getId());
+            //MNT-15498
+            if(!activitiUtil.isRetentionHistoricProcessInstanceEnabled()){
+                historyService.deleteHistoricProcessInstance(deletedInstance.getId());
+            }
             
             return result;
         }
